requirements.txt
torch
torchvision
torchaudio
numpy
tqdm
pillow
faiss-cpu
openai-clip
argparse
